Fix bug triggered by cone.
Parameters to memmove were reversed.

len is the size of the buffer. len-pos-cnt characters were copied in error
to position pos+cnt. As such this did not overflow. I.e. if len was 8
(eight chars), pos was 1 and cnt was 2, then 8-2-1=5 characters were copied
to offset 3, right at the end of the buffer. This was just plain wrong.