maildrop: rename class methods
Rename Message and Mio class methods to be the same as the logically
analogous std::streambuf methods:

Rename Message::seek to Message::pubseekpos. Mio::seek's parameters are
not what pubseekpos receives, rename its return value, keep it as a seek()
and add a compatible pubseekpos to Message.

Rename peek to sgetc.

Rename get and get_c to sbumpc.

Update the return value of pubseekpos to be consistent with std::filebuf's
behavior.