Compile mailbot.c as C++
First of a sequence of commits that convert mailbot to be C++-based.
The current mailbot.c gets renamed to mailbot-old.c. mailbot.c becomes
mailbot.C, with a minimal set of changes that are needed to compile the
code as C++.

Restructure the testsuite. Move mailbot tests into a separate testsuite
Use the same testsuite for mailbot and mailbot-old, and run them both.

mailbot-old.c will rename as is, and mailbot.C will get incremently
reimplement in C++.